Fix #18 (PHP 8: Saving of events without dates may fail)
Again, numeric calculations with empty strings; we work-around by
casting to int.

+++ includes/editevents.php
@@ -230,7 +230,7 @@ if ($dataset=='' && !$newfilename) {
     if ($datestart[$j]==$dateend[$j]) $dateend[$j] = "";
 
     //events are limited to 10 years in order to make a stop somewhere
-    if ((substr($dateend[$j],-4,4) - substr($datestart[$j],-4,4))>10) $dateend[$j] = '';
+    if (((int) substr($dateend[$j],-4,4) - (int) substr($datestart[$j],-4,4))>10) $dateend[$j] = '';
 
     //if weekly and yearly are both checked, one has to be unchecked
     if ($weekly[$j] && ($yearly[$j] || $yearly2[$j]) && $dateend[$j]) {$yearly[$j] = FALSE; $yearly2[$j] = FALSE;}
